chore: tab bar ui polish
