+++ frontend/appflowy_flutter/lib/plugins/database_view/calendar/presentation/toolbar/calendar_setting_bar.dart
@@ -23,7 +23,7 @@ class CalendarSettingBar extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      height: 40,
+      height: 20,
       child: Row(
         mainAxisAlignment: MainAxisAlignment.end,
         children: [
